+++ packages/voting-application/src/index.ts
@@ -1,3 +1,4 @@
 export * from './application-error.js';
 export * from './contracts.js';
 export * from './voting-application-service.js';
+export * from './voting-results-application-service.js';